Fix Windows Build :wrench:
Add Windows specific paths to build.rs.

Add wintypedefs.h to the project. The header files are interpreted in the same way on all platforms, so files specific to the Linux package must be added to allow for building on Windows. The alternative would be to include the complete windows.h header in the bindings, which currently fails via bindgen and produces much bigger files. See also: https://github.com/rust-lang/rust-bindgen/issues/1562